pyproject.toml
[tool.poetry]
name = "django-backend-template"
version = "0.1.0"
description = ""
authors = ["Tarun Luthra <tarunluthra987@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
black = "^21.7b0"
django-rest-framework = "^0.1.0"
python-dotenv = "^0.19.0"
django-cors-headers = "^3.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
